sdk: consolidate regtest tests; drop unused composeCommit/composeReveal
